requirements.txt
coverage==4.5.1
Django==1.11.15
djangorestframework==3.8.2
mixer==6.0.1
django-test-plus==1.1.1
